Create Generating Random nums
Hey y'all. This is a short program that demonstrates how you can generate random numbers in a given range (between 0 to 99 in this case).
I have imported 'random' module for randint() function (to generate random numbers in a range) and 'sys' module for exit() function (to exit from the program).